requirements.txt
billiard==3.6.4.0
boto3==1.10.50
botocore==1.13.50
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Flask==2.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
kombu==5.1.0
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
prompt-toolkit==3.0.20
pytz==2021.1
psycopg2-binary==2.8.5
six==1.16.0
SQLAlchemy==1.4.23
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
waitress==1.4.3
flask-jwt-extended==4.3.1
redis
